nigeria: Add Bon Maximus base, update nigeria_anambra_state
