+++ docs/spiders.rst
@@ -923,12 +923,12 @@ Nigeria
 
    scrapy crawl nigeria_abia_state
 
-.. autoclass:: kingfisher_scrapy.spiders.nigeria_anambra_state_records.NigeriaAnambraStateRecords
+.. autoclass:: kingfisher_scrapy.spiders.nigeria_anambra_state.NigeriaAnambraState
    :no-members:
 
 .. code-block:: bash
 
-   scrapy crawl nigeria_anambra_state_records
+   scrapy crawl nigeria_anambra_state
 
 .. autoclass:: kingfisher_scrapy.spiders.nigeria_budeshi_records.NigeriaBudeshiRecords
    :no-members:
