--- kingfisher_scrapy/spiders/nigeria_anambra_state_records.py
@@ -1,30 +1,0 @@
-import scrapy
-
-from kingfisher_scrapy.base_spiders import SimpleSpider
-from kingfisher_scrapy.util import components, handle_http_error
-
-
-class NigeriaAnambraStateRecords(SimpleSpider):
-    """
-    Domain
-      Anambra State
-    API documentation
-      http://www.mepbdp.org.ng/ocds/api
-    """
-    name = 'nigeria_anambra_state_records'
-
-    # SimpleSpider
-    data_type = 'record_package'
-
-    def start_requests(self):
-        yield scrapy.Request(
-            'http://www.mepbdp.org.ng/ocds/api/mda_list',
-            meta={'file_name': 'mda.json'},
-            callback=self.parse_list
-        )
-
-    @handle_http_error
-    def parse_list(self, response):
-        url_prefix = 'http://www.mepbdp.org.ng/ocds/api/mda_records/'
-        for item in response.json():
-            yield self.build_request(f'{url_prefix}{item["id"]}', formatter=components(-2))
