+++ kingfisher_scrapy/spiders/nigeria_osun_state.py
@@ -1,10 +1,11 @@
 import scrapy
 
 from kingfisher_scrapy.base_spiders import SimpleSpider
+from kingfisher_scrapy.spiders.nigeria_bon_maximus_base import NigeriaBonMaximusBase
 from kingfisher_scrapy.util import components, get_parameter_value, handle_http_error
 
 
-class NigeriaOsunState(SimpleSpider):
+class NigeriaOsunState(NigeriaBonMaximusBase):
     """
     Domain
       Nigeria Osun State Open Contracting Portal
@@ -16,17 +17,5 @@ class NigeriaOsunState(SimpleSpider):
     # SimpleSpider
     data_type = 'release_package'
 
-    # Local
+    # NigeriaBonMaximusBase
     url_prefix = 'https://egp.osunstate.gov.ng/'
-
-    def start_requests(self):
-        url = f'{self.url_prefix}awarded_contracts.php'
-        yield scrapy.Request(url, meta={'file_name': 'all.html'}, callback=self.parse_list)
-
-    @handle_http_error
-    def parse_list(self, response):
-        for url in response.xpath('//table[@id="contractTable"]/tbody/tr/td[2]/a/@href').getall():
-            # The URLs look like
-            # https://egp.osunstate.gov.ng/existing_award_details.php?id=ocds-xwwr9a-000103-OS/HLT/02
-            ocid = get_parameter_value(url, 'id').replace('/', '_')
-            yield self.build_request(f'{self.url_prefix}media/{ocid}.json', formatter=components(-1))
